pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "timekeeper-bot"
version = "0.1.0"
description = "Telegram TimeKeeper bot"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "aiogram>=3.6.0",
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy[asyncio]>=2.0.32",
  "asyncpg>=0.30.0",
  "alembic>=1.13.2",
  "redis>=5.0.8",
  "celery>=5.4.0",
  "pydantic>=2.8.2",
  "pydantic-settings>=2.4.0",
  "structlog>=24.4.0",
  "httpx>=0.27.0",
  "tenacity>=9.0.0",
  "python-dateutil>=2.9.0.post0",
  "dateparser>=1.2.0",
  "orjson>=3.10.7"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
  "pytest-asyncio>=0.24.0",
  "pytest-cov>=5.0.0",
  "mypy>=1.11.2",
  "ruff>=0.6.3",
  "pre-commit>=3.8.0",
  "types-python-dateutil>=2.9.0.20240316",
  "aiosqlite>=0.20.0"
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "ASYNC", "RUF"]
ignore = ["B008", "E501", "RUF001", "ASYNC240"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
disallow_untyped_defs = true
disallow_any_generics = true
plugins = ["pydantic.mypy"]
exclude = ["alembic"]

[[tool.mypy.overrides]]
module = ["dateparser.*", "celery.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false
check_untyped_defs = false
